feat (bindInput): allow attr/prop bindings
This enables the use of the `bindInput` directive in property and
attribute bindings.

For example:

```ts
html`
  <input type="text" .value=${bindInput(this, 'prop')}>
`;
```

This may be beneficial for SSR in particular.

Behaviour will be the same as an element binding but will delegate
setting of the element's value to lit rather than the directive dealing
with it.

Of course this won't be possible with a multi-select `<select>` element
since there is no property to bind to in that situation.